ui: combine modalLocalities and nodeCanvas
Release note: None
