+++ pkg/ui/ccl/src/views/clusterviz/containers/map/layout.spec.ts
@@ -2,7 +2,7 @@ import { assert } from "chai";
 
 import { LocalityTree } from "src/redux/localities";
 import { LocationTree } from "src/redux/locations";
-import { renderAsMap } from "./modalLocalities";
+import { renderAsMap } from "./layout";
 
 const locationTree: LocationTree = {
   zone: {
